# Rework the sort method to be more efficient for vectors.