This modifies the sort method to be much more efficient in terms of tree traversals. Swapping two items in the same focus is an expensive operation which the old sort method made extensive use of. This new method fixes that by finding the correct size of the partitions first. The downside to this method is that it makes more comparisons on average. I've noticed in my code is approximately 2x faster with this sort method.

Old sort benchmarks
```
test quicksort_reverse_sorted_list_0500 ... bench:     337,174 ns/iter (+/- 42,349)
test quicksort_reverse_sorted_list_1000 ... bench:     908,472 ns/iter (+/- 88,979)
test quicksort_reverse_sorted_list_1500 ... bench:   1,493,024 ns/iter (+/- 158,607)
test quicksort_reverse_sorted_list_2000 ... bench:   2,359,450 ns/iter (+/- 220,542)
test quicksort_reverse_sorted_list_2500 ... bench:   2,848,659 ns/iter (+/- 191,312)
test quicksort_shuffled_list_0500       ... bench:     436,730 ns/iter (+/- 41,080)
test quicksort_shuffled_list_1000       ... bench:   1,188,585 ns/iter (+/- 133,737)
test quicksort_shuffled_list_1500       ... bench:   2,052,514 ns/iter (+/- 138,074)
test quicksort_shuffled_list_2000       ... bench:   2,985,263 ns/iter (+/- 667,889)
test quicksort_shuffled_list_2500       ... bench:   3,829,387 ns/iter (+/- 318,696)
test quicksort_sorted_list_0500         ... bench:     216,016 ns/iter (+/- 22,613)
test quicksort_sorted_list_1000         ... bench:     683,743 ns/iter (+/- 46,257)
test quicksort_sorted_list_1500         ... bench:   1,625,938 ns/iter (+/- 207,219)
test quicksort_sorted_list_2000         ... bench:   1,819,486 ns/iter (+/- 257,906)
test quicksort_sorted_list_2500         ... bench:   2,782,200 ns/iter (+/- 523,779)
```

New sort benchmarks
```
test quicksort_reverse_sorted_list_0500 ... bench:     170,173 ns/iter (+/- 18,268)
test quicksort_reverse_sorted_list_1000 ... bench:     380,455 ns/iter (+/- 50,027)
test quicksort_reverse_sorted_list_1500 ... bench:     621,704 ns/iter (+/- 69,480)
test quicksort_reverse_sorted_list_2000 ... bench:     831,539 ns/iter (+/- 117,876)
test quicksort_reverse_sorted_list_2500 ... bench:   1,035,705 ns/iter (+/- 94,049)
test quicksort_shuffled_list_0500       ... bench:     201,768 ns/iter (+/- 15,222)
test quicksort_shuffled_list_1000       ... bench:     452,327 ns/iter (+/- 35,051)
test quicksort_shuffled_list_1500       ... bench:     720,080 ns/iter (+/- 96,779)
test quicksort_shuffled_list_2000       ... bench:     990,407 ns/iter (+/- 117,726)
test quicksort_shuffled_list_2500       ... bench:   1,274,549 ns/iter (+/- 99,018)
test quicksort_sorted_list_0500         ... bench:     147,221 ns/iter (+/- 19,687)
test quicksort_sorted_list_1000         ... bench:     332,869 ns/iter (+/- 32,720)
test quicksort_sorted_list_1500         ... bench:     537,311 ns/iter (+/- 42,432)
test quicksort_sorted_list_2000         ... bench:     768,849 ns/iter (+/- 80,387)
test quicksort_sorted_list_2500         ... bench:   1,042,676 ns/iter (+/- 102,850)
```