@nomad010 2020-04-19
In further testing, this PR consistently crashes when sorting lists of 5000 or more, while consistently passing when the size is 1000. I'll do more work on it and hopefully fix everything, but the PR is not ready for merge. To make it more obvious that this test fails, I'll bump the sort size up from 1000 to 5000.