@nomad010 2020-04-20
I've fixed the bug, just waiting for the checks to finish. I've tested locally on my side up to 100_000 size in the sort test, but I scaled it back to 10_000 (still 10x more than original).